+++ bloc_enhancer_test/lib/test_1/simple_bloc.dart
@@ -16,6 +16,7 @@ limitations under the License.
 */
 // --- LICENSE ---
 import 'package:bloc/bloc.dart';
+import 'package:bloc_enhancer/bloc_enhancer.dart';
 import 'package:equatable/equatable.dart';
 
 part 'simple_bloc.g.dart';